requirements.txt
pandas~=2.2.0
flask~=3.0.0
openpyxl~=3.1.2
qianfan~=0.3.7.1
python-dotenv~=1.0.0
requests~=2.31.0
flask_pymongo
bson